sharing code between fieldProjection/fieldSelection
