+++ src/powerquery-language-services/inspection/type/inspectType/common.ts
@@ -14,8 +14,8 @@ import { InspectionSettings } from "../../../inspectionSettings";
 import { inspectTypeConstant } from "./inspectTypeConstant";
 import { inspectTypeEachExpression } from "./inspectTypeEachExpression";
 import { inspectTypeErrorHandlingExpression } from "./inspectTypeErrorHandlingExpression";
-import { inspectTypeFieldProjection } from "./inspectTypeFieldProjection";
-import { inspectTypeFieldSelector } from "./inspectTypeFieldSelector";
+import { inspectTypeFieldProjection } from "./inspectTypeField/inspectTypeFieldProjection";
+import { inspectTypeFieldSelector } from "./inspectTypeField/inspectTypeFieldSelector";
 import { inspectTypeFieldSpecification } from "./inspectTypeFieldSpecification";
 import { inspectTypeFunctionExpression } from "./inspectTypeFunctionExpression";
 import { inspectTypeFunctionType } from "./inspectTypeFunctionType";
@@ -38,7 +38,7 @@ import { inspectTypeTBinOpExpression } from "./inspectTypeTBinOpExpression";
 import { inspectTypeUnaryExpression } from "./inspectTypeUnaryExpression";
 import { TypeById } from "../../typeCache";
 
-// Drop PQP.LexSettings and PQP.ParseSettings as they're not needed.
+// Drops PQP.LexSettings and PQP.ParseSettings as they're not needed.
 export interface InspectTypeState
     extends PQP.CommonSettings,
         Omit<InspectionSettings, keyof PQP.LexSettings | keyof PQP.ParseSettings> {
