+++ src/powerquery-language-services/inspection/type/inspectType/inspectTypeField/common.ts
@@ -1,0 +1,80 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { Ast, Type } from "@microsoft/powerquery-parser/lib/powerquery-parser/language";
+import {
+    NodeIdMap,
+    NodeIdMapUtils,
+    TXorNode,
+    XorNode,
+    XorNodeUtils,
+} from "@microsoft/powerquery-parser/lib/powerquery-parser/parser";
+import { Trace, TraceConstant } from "@microsoft/powerquery-parser/lib/powerquery-parser/common/trace";
+
+import { InspectTypeState, inspectXor } from "../common";
+import { LanguageServiceTraceConstant, TraceUtils } from "../../../..";
+
+// A field selection/projection is an operation a target value,
+// where the target is either an EachExpression or a RecursivePrimaryExpression.
+// In the code below that target for the FieldSelector/FieldProjection is called the scope.
+//
+// In the case of EachExpression:
+//  use whatever scope was provided in InspectionTypeState.maybeEachScopeById, else Unknown
+//
+// In the case of RecursivePrimaryExpression:
+//  the scope is the previous sibling's type, so use NodeUtils.assertGetRecursiveExpressionPreviousSibling
+
+export function inspectFieldType(state: InspectTypeState, xorNode: TXorNode): Type.TPowerQueryType {
+    const trace: Trace = state.traceManager.entry(
+        LanguageServiceTraceConstant.Type,
+        inspectFieldType.name,
+        TraceUtils.createXorNodeDetails(xorNode),
+    );
+
+    XorNodeUtils.assertIsNodeKind<Ast.FieldProjection | Ast.FieldSelector>(xorNode, [
+        Ast.NodeKind.FieldProjection,
+        Ast.NodeKind.FieldSelector,
+    ]);
+
+    // travels up the AST to see if we're in an EachExpression or RecursivePrimaryExpression.
+    const maybeEachExpression: XorNode<Ast.EachExpression> | undefined = findEachExpression(state, xorNode);
+
+    let fieldType: Type.TPowerQueryType;
+
+    // if the scope is an EachExpression
+    if (maybeEachExpression) {
+        fieldType = state.maybeEachScopeById?.get(maybeEachExpression.node.id) ?? Type.UnknownInstance;
+    }
+    // else it must be a RecursivePrimaryExpression,
+    // so grab the previous sibling of the FieldProjection/FieldSelector
+    else {
+        const previousSibling: TXorNode = NodeIdMapUtils.assertGetRecursiveExpressionPreviousSibling(
+            state.nodeIdMapCollection,
+            xorNode.node.id,
+        );
+
+        fieldType = inspectXor(state, previousSibling);
+    }
+
+    trace.exit({ [TraceConstant.Result]: TraceUtils.createTypeDetails(fieldType) });
+
+    return fieldType;
+}
+
+function findEachExpression(state: InspectTypeState, xorNode: TXorNode): XorNode<Ast.EachExpression> | undefined {
+    const nodeIdMapCollection: NodeIdMap.Collection = state.nodeIdMapCollection;
+    let maybeParent: TXorNode | undefined = NodeIdMapUtils.maybeParentXor(nodeIdMapCollection, xorNode.node.id);
+
+    while (maybeParent) {
+        switch (maybeParent.node.kind) {
+            case Ast.NodeKind.EachExpression:
+                return maybeParent as XorNode<Ast.EachExpression>;
+
+            default:
+                maybeParent = NodeIdMapUtils.maybeParentXor(nodeIdMapCollection, maybeParent.node.id);
+                break;
+        }
+    }
+
+    return undefined;
+}
