+++ src/powerquery-language-services/inspection/type/inspectType/inspectTypeField/inspectTypeFieldSelector.ts
@@ -2,21 +2,16 @@
 // Licensed under the MIT license.
 
 import { Ast, Type } from "@microsoft/powerquery-parser/lib/powerquery-parser/language";
-import {
-    NodeIdMap,
-    NodeIdMapUtils,
-    TXorNode,
-    XorNode,
-    XorNodeUtils,
-} from "@microsoft/powerquery-parser/lib/powerquery-parser/parser";
+import { NodeIdMapUtils, TXorNode, XorNodeUtils } from "@microsoft/powerquery-parser/lib/powerquery-parser/parser";
 import { Trace, TraceConstant } from "@microsoft/powerquery-parser/lib/powerquery-parser/common/trace";
 
-import { InspectTypeState, inspectXor } from "./common";
-import { LanguageServiceTraceConstant, TraceUtils } from "../../..";
+import { LanguageServiceTraceConstant, TraceUtils } from "../../../..";
+import { inspectFieldType } from "./common";
+import { InspectTypeState } from "../common";
 
-// A field selector is an operation done on some value.
+// A field selection/projection is an operation done on some value.
 // The target can be either an EachExpression or a RecursivePrimaryExpression.
-// In the code below that target for the FieldSelector is called the scope.
+// In the code below that target for the FieldSelector/FieldProjection is called the scope.
 //
 // In the case of EachExpression:
 //  use whatever scope was provided in InspectionTypeState.maybeEachScopeById, else Unknown
@@ -47,20 +42,7 @@ export function inspectTypeFieldSelector(state: InspectTypeState, xorNode: TXorN
     }
 
     const fieldName: string = maybeFieldName.literal;
-
-    // travels up the AST to see if we're in an EachExpression or RecursivePrimaryExpression.
-    const maybeEachExpression: XorNode<Ast.EachExpression> | undefined = findEachExpression(state, xorNode);
-    let fieldType: Type.TPowerQueryType;
-
-    // if the scope is an EachExpression
-    if (maybeEachExpression) {
-        fieldType = state.maybeEachScopeById?.get(maybeEachExpression.node.id) ?? Type.UnknownInstance;
-    }
-    // else it must be a RecursivePrimaryExpression,
-    // so grab the previous sibling of the FieldSelector
-    else {
-        fieldType = inspectRecursivePrimaryExpressionFieldSelector(state, xorNode);
-    }
+    const fieldType: Type.TPowerQueryType = inspectFieldType(state, xorNode);
 
     const isOptional: boolean =
         NodeIdMapUtils.maybeUnboxNthChildIfAstChecked(
@@ -74,24 +56,6 @@ export function inspectTypeFieldSelector(state: InspectTypeState, xorNode: TXorN
     trace.exit({ [TraceConstant.Result]: TraceUtils.createTypeDetails(result) });
 
     return result;
-}
-
-function findEachExpression(state: InspectTypeState, xorNode: TXorNode): XorNode<Ast.EachExpression> | undefined {
-    const nodeIdMapCollection: NodeIdMap.Collection = state.nodeIdMapCollection;
-    let maybeParent: TXorNode | undefined = NodeIdMapUtils.maybeParentXor(nodeIdMapCollection, xorNode.node.id);
-
-    while (maybeParent) {
-        switch (maybeParent.node.kind) {
-            case Ast.NodeKind.EachExpression:
-                return maybeParent as XorNode<Ast.EachExpression>;
-
-            default:
-                maybeParent = NodeIdMapUtils.maybeParentXor(nodeIdMapCollection, maybeParent.node.id);
-                break;
-        }
-    }
-
-    return undefined;
 }
 
 function getFieldSelectorType(
@@ -133,16 +97,4 @@ function inspectRecordOrTable(
     } else {
         return isOptional ? Type.NullInstance : Type.NoneInstance;
     }
-}
-
-function inspectRecursivePrimaryExpressionFieldSelector(
-    state: InspectTypeState,
-    xorNode: TXorNode,
-): Type.TPowerQueryType {
-    const previousSibling: TXorNode = NodeIdMapUtils.assertGetRecursiveExpressionPreviousSibling(
-        state.nodeIdMapCollection,
-        xorNode.node.id,
-    );
-
-    return inspectXor(state, previousSibling);
 }
